Moved new functions to downloader
